third_party/WebKit/LayoutTests/animations/prefixed/string-keyframes-identifier-prefixed.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
@-webkit-keyframes "string id" {
  from { background: green; }
  to { background: green; }
}
#target {
  background: red;
  -webkit-animation: "string id" 1s forwards;
}
</style>
<div id="target"></div>
<script>
test(function() {
  assert_equals(getComputedStyle(target).backgroundColor, 'rgb(0, 128, 0)');
}, 'Check that strings are permitted as @-webkit-keyframes identifiers');
</script>
